feat: update order management, enhance order creation and retrieval, and modify order item handling aligned to new DB schema

+++ src/main/java/edu/icet/ecom/dto/AssignWaiterRequest.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AssignWaiterRequest {
-    private Long orderId;
+    private Integer orderId;
     private Long waiterId;
 }
