+++ src/main/java/edu/icet/ecom/repository/impl/OrderAssignmentRepositoryImpl.java
@@ -17,7 +17,7 @@ public class OrderAssignmentRepositoryImpl implements OrderAssignmentRepository 
     }
 
     @Override
-    public void assignWaiter(Long orderId, Long waiterId) {
+    public void assignWaiter(Integer orderId, Long waiterId) {
         String sql = """
                 INSERT INTO order_assignments(order_id, waiter_id)
                 VALUES (?,?)
